+++ docker/dc.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -ex
+
+# Alias for docker-compose
+
+docker-compose $@